learned that should only be in the parent component
In this PR, I raised the state to be in the parent App component, so that NavBar, Counter, and Counters components can all benefit from the state
