+++ src/components/navbar.jsx
@@ -1,0 +1,18 @@
+import React, { Component } from "react";
+class NavBar extends Component {
+
+  render() {
+    return (
+      <nav className="navbar navbar-light bg-light">
+        <p className="navbar-brand">
+          Navbar {" "}
+          <span className="badge badge-pill badge-secondary">
+            {this.props.totalCounters}
+          </span>
+        </p>
+      </nav>
+    );
+  }
+}
+
+export default NavBar;
